requirements.txt
mcp[cli]
python-dotenv
sqlite-vec
torch
transformers
jira
